requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart
websockets
aiofiles

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# ML / Analysis
torch>=2.1.0
open-clip-torch>=2.24.0

# Translation
deep-translator>=1.11.0

# Colab tunneling (optional)
# pyngrok
